Bug 801383 Fix fallout from bug 779473 - make nsresult an enum class. Fix most of mailnews. Patch started by mconley, completed by irving. r=Standard8

+++ mailnews/addrbook/src/nsAbLDAPAutoCompFormatter.cpp
@@ -219,7 +219,7 @@ nsAbLDAPAutoCompFormatter::FormatException(int32_t aState,
         // put the entire nsresult in string form
         //
         errCodeNum.AppendLiteral("0x");
-        errCodeNum.AppendInt(aErrorCode, 16);    
+        errCodeNum.AppendInt(static_cast<uint32_t>(aErrorCode), 16);
 
         // figure out the key to index into the string bundle
         //
